modularized and refractored all code
