--- rplh_decentric/response_model.py
@@ -1,7 +1,0 @@
-from typing import Dict, List
-from pydantic import BaseModel
-
-class LocalAgent(BaseModel):
-    attitude: List[str]
-    reasoning: str
-    actions_plan: Dict
